data/requirements.txt
python-chess==0.30.1
tqdm==4.42.0
